requirements.txt
torch
torchvision==0.8.2
timm==0.3.2
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
deepspeed==0.4.0
scipy
torchmetrics
